Tier 1: MAM economy bundle (MouseSearch port)
Adds four user-visible features to Seshat, all gated behind new
DEFAULT_SETTINGS flags that default to False — first deploy is a
silent no-op until the user opts in per-feature via the new
MamPage sections.

  1. VIP auto-buy — scheduled BP→VIP-time purchase on a
     settings-driven interval, with min-bonus floor and
     choose-your-weeks dropdown (4/8/12/max).
  2. Upload-credit auto-buy — three independent triggers
     (ratio / buffer / bonus-excess) evaluated in priority
     order, with fixed-chunk or excess-derived buy amounts.
  3. Pre-download buffer gate — new policy-engine step that
     skips a grab (both IRC autograb AND manual inject) when
     the torrent size + safety margin would overflow the user's
     upload buffer. Throttled ntfy notification on first block
     per 6h window per trigger type.
  4. Per-grab wedge + personal-FL offers — optional checkboxes
     on the BookSidebar "Send to pipeline" flow that drain a
     wedge from the pool OR spend 50k BP for spendtype=personalFL
     before the grab runs.

New API surface under /api/v1/mam/economy:

  - config (GET/PUT) — whitelisted config merge
  - vip/buy, upload/buy, personal-fl/buy — manual buys
  - audit — MamPage history tile
  - preflight — BufferInsufficientBanner one-click retry math

New backend modules:
  app/mam/bonus_buy.py           — bonusBuy.php wrappers
  app/mam/economy.py             — pure decision engine
  app/orchestrator/economy_scheduler.py — two supervised loops
  app/routers/economy.py         — HTTP surface
  app/storage/economy_audit.py   — one-row-per-decision audit

New frontend:
  frontend/src/lib/economyApi.ts
  frontend/src/components/BufferInsufficientBanner.tsx
  + three new Sections in MamPage.tsx
  + per-grab offer controls in BookSidebar.tsx

Schema: new economy_audit table (SCHEMA + 3 MIGRATIONS entries).

UserStatus.seedbonus widened int -> float (bonusBuy.php returns
fractional values); UserStatus gains upload_buffer_bytes field
with a uploaded-downloaded fallback when MAM omits upload_buffer.

Test coverage: 142 new tests; full suite 1016 passed, 7 skipped.

Feature attribution: the economy feature set is inspired by
MouseSearch (github.com/sevenlayercookie/MouseSearch) — a public
Quart/Bootstrap MAM app. Port was done under "practical
independence": read the public source to clarify behavior
(notably the personalFL URL shape, where MAM wants the epoch-ms
cache-buster duplicated as trailing path segment AND timestamp
query param), but every line below was written from behavioral
notes and the MAM API shape, not transcribed.

Out of scope for Tier 1 (shipping separately):
  - SSE live torrent polling (Tier 2)
  - PWA (Tier 3)
  - Details UI polish (Phase 4)
  - Template path extension, README docs (Phases 5-6)

Test plan for manual UAT is in TIER1_TEST_PLAN.md.

+++ app/routers/mam.py
@@ -52,7 +52,7 @@ class MamStatusResponse(BaseModel):
     classname: Optional[str] = None
     ratio: Optional[float] = None
     wedges: Optional[int] = None
-    seedbonus: Optional[int] = None
+    seedbonus: Optional[float] = None
     uploaded_bytes: Optional[int] = None
     downloaded_bytes: Optional[int] = None
     error: Optional[str] = None
